Simplify choiceNodes' init-process.
ChoiceNodes' need to set their choices' parent to their own parent (and
some more fields, ie. act as if the choice is actually at the
choiceNodes' place). This isn't possible during construction as the
parent is unknown at that time.
Using __index and an intermediary metatable, it is possible to (maybe
even more clearly) represent this: the __index redirects keys that were
not found in the choice to the choiceNode during runtime (the nodes' own
metatable has priority, of course).