Use tokenizer to create helper-snippet for regex-triggers.
This is mainly useful for completion-engines, they could, up till now,
only insert the trigger, which is likely to be invalid for
regTrig-snippets (completion inserts "head%d", cannot be expanded. With
this, it would _expand_ "head${1:%d}$0", the cursor would be behind the
snippet after completion, the helper-snippet is unlinked and the user
may `expand` and get their regex-snippet).
`snip:get_pattern_expand_helper():trigger_expand(
    Luasnip_current_nodes[vim.api.nvim_get_current_buf])`
would expand the helper-snippet.