fix(insertNode): node ending with \n leads to wrong selection.
the extmark is one behind the last character of the selection, if the last
char is a linebreak, the extmark is placed at column 0 of the next line.

Previously this lead to the movement `<some lnum>G-1lzv`, (`-` is one line
up), this patch fixes the check for the `regular` case and implements
the special case for end-on-\n:
(for pos `{n, 0}`) <n-1>G$zv (NOTE: This might fail with settings that change
the behaviour of $) eg. move onto the \n of the previous line.

(Also added: open folds in all cases.)