Allow the default for a given key to be passed to the constructor.
If the default is specified more than once it's undefined which one is
actually used.
If the default is set within a dynamicNode, it will not influence
restoreNodes outside the dynamicNode directly after expansion. The
reason for this is that dynamicNodes can only be constructed (and their
subsnip_init() called) after put_initial() has been called for all of
the argnodes (actually all the nodes, with more effort this could be
optimized to just the argnodes).