fix: fix+improve update-behaviour of restore/dynamic/snippetNode.
Simply copying `dependents` isn't enough: as nodes may be deeply nested,
the dependents of the parent might change, at which point all the
dependents of children several layers deeper would also have to be
updated.
Overriding the update_dependents to include an update to the parent is
simpler+may be useful later should the argnode-system be rewritten.