fix: Perform snip-delete-check before updating.
If they are registered on the same autocommand they are guaranteed to be
executed in the same order they were defined.
If the update is performed before delete, a snippet that no longer has
any text may update it's function/dynamicNode, making it contain text again,
so it wouldn't be deleted by the subsequent delete-check.