perf: improve primitive algorithm for adjusting marks inside snippet.
Previously all gravities were adjusted, meaning they may have to be adjusted
back right away (at least if this was called as part of the
`self.parent:enter_node(self.indx)` from snippet:enter_node(indx)).

This change leads to a noticeable performance-improvement if the jump
key is held down and the snippet has a large number of nodes.