Don't refrain from updating if the text still matches.
A node will only update if the args no longer match either way, and this
could prevent updates if eg. two functionNodes depend on one insertNode,
only one of which is visible at expansion. As soon as the other becomes
visible, it should update, but won't because the text in the insertNode
hasn't changed, therefore update isn't triggered.