Set active to false by default.
Can cause problems if the dynamicNode is a direct child of a choiceNode,
as they get the choiceNode as their __index, so `active`, if nil, could
be set to true despite the node not being entered yet (as the choiceNode
is entered).
This is, for now, prevented by simply setting `active`, but it's
questionable whether it really makes sense redirect nil-values to the
choiceNode (the main reason for it is that choices should act as if they
are directly inside the parent of the choiceNode (at the place of the
choiceNode), which is easier if they just get some values from the
choiceNode if they are not set on the choice).