Add concept of static visibility.
Normally this snippet:
s("trig", {
	i(1, "cccc"),
	t" ",
	c(2, {
		t"aaaa",
		i(nil, "bbbb")
	}),
	f(function(args) return args[1][1]..args[2][1] end, {ai[2][2], 1} )
})

would have static text {"cccc aaaabbbbcccc"}, which is wrong, as the
functionNode should not be able to be evaluated if not all argnodes are
visible.
With this change, it's possible to track which nodes are visible (and therefore
may provide an arg to a function/dynamicNode) while evaluating static text or
a docstring.