Add static versions of all the update-methods.
These are required to correctly generate static text if an argnode is
generated by a dynamicNode: with the previous system, the snips are
generated during get_static_text(), but if not all argnodes are present,
the generation will fail, causing incorrect text for the dynamicNode.

With these changes, the updating will be performed recursively (very similar
to usual expansion),which means that dynamic/functionNodes can be updated by
newly generated argnodes. This means that in the end, all possible, the
function might rely on being evaluated in some specific text-area, snippets
will be generated.