feat: don't chage mode unnecessarily.
Previously, the mode was always changed to NORMAL before the cursor is
moved in any way.
Now a check is performed, if current- and target-mode are both INSERT there
is no need to switch to NORMAL as the cursor can just be moved via <Up>,
<Down>, ....
There still is a switch to NORMAL for SELECTing insertNodes, but this is
a **noticeable** improvement when it comes to jumping between nodes.