Breaking Change: split ext_opts into node_ext_opts and child_ext_opts.
The former is applied to this specific node, the latter to children of
that snippetNode.
Also add merge_{node,child}_ext_opts for control merge-behaviour of eg.
a nodes ext_opts with that of the surrounding snippet (both default to
true).
This is a breaking change because function/dynamicNode no longer have
the variable number of user_args as their last arg. That is now a table,
`opts`, which may contain the `user_args`.

```lua
d(pos, func, argnodes, user_arg1, user_arg2)

-- becomes

d(pos, func, argnodes, {user_args = {user_arg1, user_arg2}})

-- and likewise for functionNode.
```