Tweak DOC.md to make panvimdoc happy
I'm guessing that panvimdoc (or pandoc?) doesn't like `*` to introduce
lists. In any case, the formatting of the list under the BASICS heading
was lost. I've switched this list to use `-` (which you use elsewhere)
and I've indented the sub-list.

There are actually a bunch of places where the translation from markdown
to vim help text is less than ideal, but this seemed like low hanging
fruit. I am happy to try to make other tweaks (e.g., lots of whitespace
changes) if you're open to it. Just let me know.

Thanks for LuaSnip. I'm just getting to know it, but I'm enjoying it so
far.