snip_expand: delete clear_region before expanding.
If it is deleted after expanding, there are difficulties getting the
correct indent (for example, if the trigger changes the indent (it might
match the entire line...), this is not reflected in the snippets
indent).
This could be fixed by
A: still deleting the trigger after expanding, but adjusting the line
before the trigger with clear_region (remove parts removed by
clear_region) or
B: populate env before removing clear_region (to get correct
$TM_CURRENT_LINE), removing clear_region (expand-position has to stay
the same relative to clear_region, this is for now done by setting an
extmark on it before clearing and retrieving its position afterward).

B feels cleaner, mainly because it also removes the trigger before
expanding, which just seems right.
Having to remember the expand-position via extmark is a bit unfortunate,
but alright.