fix(f/dNode): always add potential argnodes to dependents_dict.
Situation: snippet with fNode nested inside dNode, but the fNode depends
on an iNode outside the dNode.
In this case, set_argnodes has to be called for the iNode after
fNode:set_dependents, but that never happens.
So, instead, we relinquish the requirement that the insert-nodes are
only added to the dependents_dict if they are actually, surely, an
argnode (because they are already listed inside it), and just add them
preemptively.