f/dNode: if arg is node, store it instead of its position in dependents_dict.
If it is not in the snippet at set_dependents (which may just be called
once), it is not stored in the dependents_dict at all. Since argnodes
need to notify their dependents when they were changed, they have to be,
though. So, for nodes as arg, we can just store the node itself in
dependents_dict, obviously the lookup-routine needs to be modified a
bit, but this seems to be an okay solution.