add snippetType as option on snippet creation
snippetType shall be "autosnippet" or "snippet" (or nil -> unset -> take
default value from add_snippets). Note: snippetType is used for indexing
(it already is a string, which should be interned by lua)

First draft can be seen at
[this issue](https://github.com/L3MON4D3/LuaSnip/issues/482#issuecomment-1207361553),
adapted because `selfpopulating tables` were merged in the meantime and
the option was renamed.

For a dicussion on naming the new option key see
[the comments here](https://github.com/L3MON4D3/LuaSnip/pull/513#issuecomment-1245778674)