feat! implement condition objects which implement various operators + tests
Works by
- providing a factory `make_condition` which wraps functions as
  condition objects
- providing two modules `conditions.show` and `condition.expand` which
  contain a collection of common conditions already wrapped as condition
  objects

Weird might be the decision to use `%` as `==`-operator. This way chosen
because we can`t use `__eq` as this automatically converts the return
value to a boolean (we need to return condition objects). So we decided
to use something which make ones head scratch and look it up in the docs
(or simply don't use it)

Note: one still can continue to use the `expand_conditions`
module (just rediects to the `conditions/expand` module) for backwards
compatibility. This is deprecated though.