Add path to luasnips loader
In order to keep my snippets in better order, I wanted to split my
snippets into multiple files (one snippet per file), as here:

```
~/.config/nvim/snippets
├── all.lua
├── ...
└── lua
    ├── init.lua
    ├── snippet_1.lua
    ├── snippet_2.lua
    └── snippet_3.lua
```

This means having a `snippets/lua/init.lua` file that imports the other
files in the directory, e.g.

```lua
local snippets = {
    require('lua.snippet_1'),
    require('lua.snippet_2'),
    require('lua.snippet_3'),
}

return snippets, {}
```

The problem that I ran into was that the require function didn't work as
it didn't have an appropriate path property, so could not resolve the
imports.

In order to resolve this issue, I have implemented a new property on the
config, `snip_path`, which takes a lua path string.

If this `snip_path` property exists it is then injected into the snippet
(at the same time as the env is injected), changing the `package.path`
property which is used in Neovim to resolve imports.