perf: make some tables lazily-initialized.
This reduces the upfront cost of constructing them, and instead does so
on-demand.

test-change: since earlier, lots of stuff was loaded in advance, this
"replace the entire vim-table" worked. Now, with the lazy tables, some
required functions cannot be resolved, and we have to be a bit more
clever here.