don't modify jumplist-insertion too much, does not match some assumptions.
In particular, snip.next.next of a i(0), or snip.prev.prev of a i(-1) is
often assumed to be the next/previous snippet, which is not the case if
we do the modifications from the parent.

Instead: just modify behaviour of i(-1), such that it is just ignored
while jumping.