improve adjustment of extmark-rgravs.
Both performance-wise, by eliding many, many adjustments, and
implementation-wise.

For example, as soon as a parent does not share an endpoint with the
node we want to focus, we don't have to touch its rgravs, since they can
only be pushed away from the inserted text, no matter their gravity.