overhaul snippet-updates.
Previously:
InsertNodes trigger an update on input_leave.

This works, but, if updates invalidate nodes that are currently involved
in being jumped over, we'd have to abort and roll back the jump, or do
some other recovery.

To avoid this, update_dependents is done _before_ any jumping is
performed (this is really more elegant now since we can keep track of
which nodes are "above" some insertNode, and then just get all of them
and their dependents (see
node_util.find_node_dependents/collect_dependents))

So, now `ls.jump` does essentially:
* store position of cursor relative to active node
* collect nodes that depend on the text of this node (so, all dependents
  of all parents of this node)
* update them
* try to find an equivalent node (node with the same key, or if a
  restoreNode is present, the exact same node :D) and perform the jump
  from it
* if an equivalent node could not be found, just enter the first
  dynamicNode (starting from the previously active node) and enter it

Obviously, this only really works well if an equivalent of the current
node can be found in the new generated nodes.

Similarly, active_update_dependents