feat: ensure strict ordering of feedkeys initiated by luasnip.
This allows us to first of all, be certain that successive ls.jump()
always end up at the correct position, but also enables running all
these feedkeys with "ni", which means that they will be executed at the
correct time even if the typeahead contains other keys! This makes
luasnip behave deterministically in macros, which was certainly not the
case before!

We can also finally remove all the ugly vim.wait(...)-calls from the
tests :D (at least those related to waiting on luasnip-actions to
complete, we still need some related to filesystem-events.)
Now, these only have to happen for actions that are not hooked into the
feedkeys-action-queue, like `ls.expand`. IMO while not perfect, this is
a great improvement because `ls.expand` are rarely called back-to-back,
while `ls.jump` are.

A last neat addition is that autosnippets can be expanded properly when
replaying a macro by inserting expand_auto before other typeahead, thus
making it take place at the correct position.