store cursor-position in snippetString to more accurately restore it.
Cache `static_text` (for insertNodes) and don't query it anew while
luasnip is operating. This is valid under the assumption that all
changes to the buffer are due to luasnip while an api-function (jump,
expand, etc.) is running.

This is enabled by session.luasnip_changedtick
which is set as soon as an api-function is called, and prevents
re-fetching snippetStrings from the buffer (which in turn allows us to
set the cursor-position once, and then have it propagate through all
updates that are triggered subsequently).

This commit also replaces no_region_check_wrap with api_do, which is
more general (handles both jump_active, which prevents recursive
api-calls and luasnip_changedtick)