update after snip_expand.
Has to happen because we modified text.
I don't like using vim.schedule here, but it seems like this is the
only way of getting the desired behaviour into all possible ways of
expanding snippets (direct snip_expand is used by cmp_luasnip, so can't
just handle `expand` and its variants in init.lua (or, we could do so
and submit a PR to cmp_luasnip, but let's wait with that until this
actually becomes problematic, which I don't think it will)).