rip out update_depth again.
I don't think that this is a good idea, at least currently, where I
don't yet have a good feel for the way self-dependent dynamicNodes are
going to be used.

Difficulties: picking a good default depth. Even something as outwardly
innocuous as 100 is bad when applied to a snippet that doubles some
character. On the other hand, this can totally kill "chains" of updates,
where some inner snippet updates text that is suddenly changed by an
outer snippet...

So, for now, I'm putting this down as user error, and even though having
to kill your nvim-instance because it is caught in an infinite loop is
not great at all, it shouldn't occur to often, and all ways around this
I can think of may impede absolutely valid usage of dynamicNodes.