Update README with BSD/GNU Makefile distinction
Added a little call-out in the readme for users regarding GNU make installation.

This has bitten me a few times and it took a moment to remember what to do. I have no opinions on whatever GNU/BSD makefile choices are made in the project, just that it's documented so I can reference it later.

Thanks.